Babulal and PC
Students of IIIT-Delhi think a lot about the world so they switch off
the already ON PCs
while leaving the lab. It is morning in IIIT-Delhi and all the PCs of
the lab are turned off.
The guard of IIIT-Delhi Mr.Babulal is a weird fellow. He does a weird
thing every
morning. If there are n PCs in the lab then he takes n rounds back and
forth and in the
ith round toggles the state of the PCs whose index are not divisible by
i and then comes
back to the 1st PC without turning ON/OFF any PC.
You have to tell the state of the kth PC when Babulal has completed all
his rounds.
Assume all PCs are in order and are indexed from 1 to N (inclusive) and
Babulal starts
from 1st PC. Input will satisfy all constraints given.
Input
First Line contains an integer T, the number of test cases.
T lines follow each with two space separated integers N and k
Output
T lines with “off” or “on”. Output “on” if the kth pc is turned on and
“off” if kth pc is off
(without the quotation marks).
Constraints
1 <= T <= 50000
1 <= k <= N <= 10^9
Time Limit: 1sec
Sample Input
2
5 4
5 2
Sample Output
off
on